api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "eliabevces",email = "eliabevces@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "redis (>=6.1.0,<7.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "blurhash-python (>=1.2.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
